crawler/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tqdm==4.66.1